+++ src/bimark.ts
@@ -1,74 +1,27 @@
 import { remark } from "remark";
 import { visit } from "unist-util-visit";
-import uslug from "uslug";
-import {
-  DefIdGenerator,
-  Definition,
-  Fragment,
-  Position,
-  RefIdGenerator,
-} from "./model";
+import { BiDoc } from "./bidoc";
+import { Definition, Fragment, Position, RefIdGenerator } from "./model";
 import { BiParser } from "./parser";
 
-export class BiMark {
-  /** name/alias => Definition */
-  readonly name2def: Map<string, Definition>;
-  /** id => Definition */
-  readonly id2def: Map<string, Definition>;
-  readonly defIdGenerator: DefIdGenerator;
-  readonly refIdGenerator: RefIdGenerator;
-
+export class BiMark extends BiDoc {
   constructor(
     options?: Partial<Pick<BiMark, "refIdGenerator" | "defIdGenerator">>
   ) {
-    this.defIdGenerator = options?.defIdGenerator ?? ((name) => uslug(name));
-    this.refIdGenerator =
-      options?.refIdGenerator ??
-      ((_, def, index) => `${def.id}-ref-${index + 1}`);
-
-    this.name2def = new Map();
-    this.id2def = new Map();
+    super(options);
   }
 
-  /**
-   * Collect definitions from a markdown document.
-   */
-  collect(path: string, md: string) {
-    const ast = remark.parse(md);
+  protected parseTextNode(
+    content: string,
+    cb: (text: string, position: Position) => void
+  ): void {
+    const ast = remark.parse(content);
     visit(ast, (node) => {
-      if (node.type == "text") {
-        BiParser.collectDefinition(
-          node.value,
-          path,
-          node.position!,
-          this.defIdGenerator
-        ).defs.forEach((d) => {
-          // check name/alias/id duplication
-          if (this.name2def.has(d.name))
-            throw new Error(
-              `Duplicate definition name: ${d.name} in file ${path}`
-            );
-          if (this.id2def.has(d.id))
-            throw new Error(`Duplicate definition id: ${d.id} in file ${path}`);
-          d.alias.forEach((a) => {
-            if (this.name2def.has(a))
-              throw new Error(
-                `Duplicate definition name: ${a} in file ${path}`
-              );
-          });
-
-          this.name2def.set(d.name, d);
-          this.id2def.set(d.id, d);
-          d.alias.forEach((a) => {
-            this.name2def.set(a, d);
-          });
-        });
-      }
+      if (node.type == "text") cb(node.value, node.position!);
     });
-    return this;
   }
 
-  private renderDefinition(
+  protected renderDefinition(
     path: string,
     fragments: Fragment[],
     options: { showBrackets: boolean; showAlias: boolean }
@@ -92,7 +45,7 @@ export class BiMark {
     return res.fragments;
   }
 
-  private renderExplicitOrEscapedReference(
+  protected renderExplicitOrEscapedReference(
     path: string,
     fragments: Fragment[],
     options: { showBrackets: boolean; html: boolean }
@@ -139,7 +92,7 @@ export class BiMark {
     );
   }
 
-  private renderImplicitReference(
+  protected renderImplicitReference(
     path: string,
     fragments: Fragment[],
     def: Definition,
@@ -169,36 +122,6 @@ export class BiMark {
         };
       }
     );
-  }
-
-  private renderText(
-    path: string,
-    s: string,
-    position: Position,
-    options: {
-      def: { showAlias: boolean; showBrackets: boolean };
-      ref: { showBrackets: boolean; html: boolean };
-    }
-  ) {
-    let fragments: Fragment[] = [{ content: s, skip: false, position }];
-
-    fragments = this.renderDefinition(path, fragments, {
-      showAlias: options.def.showAlias,
-      showBrackets: options.def.showBrackets,
-    });
-    fragments = this.renderExplicitOrEscapedReference(path, fragments, {
-      showBrackets: options.ref.showBrackets,
-      html: options.ref.html,
-    });
-
-    this.name2def.forEach((def, content) => {
-      fragments = this.renderImplicitReference(path, fragments, def, content, {
-        showBrackets: options.ref.showBrackets,
-        html: options.ref.html,
-      });
-    });
-
-    return fragments.map((f) => f.content).join("");
   }
 
   /**
@@ -252,21 +175,5 @@ export class BiMark {
   ) {
     const path = options?.path ?? "";
     return new BiMark().collect(path, md).render(path, md, options);
-  }
-
-  /**
-   * Get the references of a definition.
-   * Return an array of link address to the references.
-   */
-  getReverseRefs(options: { id: string } | { name: string }) {
-    const def =
-      "id" in options
-        ? this.id2def.get(options.id)
-        : this.name2def.get(options.name);
-
-    if (!def)
-      throw new Error(`Definition not found: ${JSON.stringify(options)}`);
-
-    return def.refs.map((p, i) => `${p}#${this.refIdGenerator(p, def, i)}`);
   }
 }